Optimize the logic related to table structure upgrade.(https://github.com/dora4/dcache-android/issues/15)

+++ dcache/src/main/java/dora/db/dao/OrmDao.kt
@@ -15,14 +15,15 @@ import dora.db.constraint.Id
 import dora.db.constraint.PrimaryKey
 import dora.db.converter.PropertyConverter
 import dora.db.exception.OrmResultCreationException
+import dora.db.exception.OrmStateException
 import dora.db.exception.UnsupportedDataTypeException
 import dora.db.table.*
 import java.lang.reflect.*
 import java.util.*
 
-class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>) : Dao<T> {
-
-    private val database: SQLiteDatabase = Orm.getDB()
+class OrmDao<T : OrmTable> internal @JvmOverloads constructor(val beanClass: Class<T>,
+                                                val database: SQLiteDatabase? = if (Orm.isPrepared())
+                                                    Orm.getDB() else null) : Dao<T> {
 
     private fun isAssignableFromBoolean(fieldType: Class<*>): Boolean {
         return Boolean::class.javaPrimitiveType!!.isAssignableFrom(fieldType) ||
@@ -253,10 +254,16 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
         }
 
     override fun insert(bean: T): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         return insertInternal(bean, database)
     }
 
     override fun insert(beans: List<T>): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         return insertInternal(beans, database)
     }
 
@@ -278,10 +285,16 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun delete(builder: WhereBuilder): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         return deleteInternal(builder, database)
     }
 
     override fun delete(bean: T): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val field = bean.javaClass.getDeclaredField(getPrimaryKeyFieldName(bean))
         field.isAccessible = true
         val name = TableManager.getColumnName(field)
@@ -290,6 +303,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun deleteAll(): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         return deleteAllInternal(database)
     }
 
@@ -348,10 +364,16 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun update(builder: WhereBuilder, newBean: T): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         return updateInternal(builder, newBean, database)
     }
 
     override fun update(bean: T): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val field = bean.javaClass.getDeclaredField(getPrimaryKeyFieldName(bean))
         field.isAccessible = true
         val name = TableManager.getColumnName(field)
@@ -377,6 +399,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun selectAllInternal(): List<T> {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val tableName: String = TableManager.getTableName(beanClass)
         val cursor = database.query(tableName, null, null, null, null, null, null)
         return getResult(cursor)
@@ -391,6 +416,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun selectInternal(builder: QueryBuilder): List<T> {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val tableName: String = TableManager.getTableName(beanClass)
         val columns: Array<String>? = builder.getColumns()
         val group: String = builder.getGroup()
@@ -409,6 +437,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun selectOneInternal(): T? {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val tableName: String = TableManager.getTableName(beanClass)
         val cursor = database.query(tableName, null, null, null,
             null, null, null)
@@ -429,6 +460,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun selectOneInternal(builder: QueryBuilder): T? {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val beans: List<T> = select(builder)
         if (beans.isNotEmpty()) {
             return beans[0]
@@ -485,6 +519,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun countInternal() : Long {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         var count: Long = 0
         try {
             val tableName: String = TableManager.getTableName(beanClass)
@@ -509,6 +546,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun addColumn(fieldName: String): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val tableName = TableManager.getTableName(beanClass)
         try {
             val field = beanClass.getDeclaredField(fieldName)
@@ -533,6 +573,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun renameColumn(fieldName: String, oldColumnName: String): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         val tableName = TableManager.getTableName(beanClass)
         try {
             val field = beanClass.getDeclaredField(fieldName)
@@ -558,6 +601,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     override fun drop(): Boolean {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         try {
             val tableName = TableManager.getTableName(beanClass)
             val sql = TableManager.DROP_TABLE + TableManager.SPACE + tableName
@@ -571,6 +617,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
     }
 
     private fun countInternal(builder: QueryBuilder) : Long {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         var count: Long = 0
         try {
             val tableName: String = TableManager.getTableName(beanClass)
@@ -849,6 +898,9 @@ class OrmDao<T : OrmTable> internal constructor(private val beanClass: Class<T>)
      * @see dora.db.Transaction
      */
     fun runInTransaction(block:() -> T) {
+        if (database == null) {
+            throw OrmStateException("Database is not exists.")
+        }
         try {
             // Begin the transaction.
             // 简体中文：开始事务
